converted Faq into Component
