+++ app/components/header.tsx
@@ -1,6 +1,7 @@
 /* This example requires Tailwind CSS v2.0+ */
 import { Fragment } from 'react';
 import { Popover, Transition } from '@headlessui/react';
+import { Container ,Typography} from '@mui/material';
 import {
   BookmarkAltIcon,
   CalendarIcon,
